GLUE-381 CustomersRestApi multiple bugfixes

+++ Bundles/CustomersRestApi/src/Spryker/Glue/CustomersRestApi/Processor/Addresses/AddressesWriter.php
@@ -7,13 +7,14 @@
 
 namespace Spryker\Glue\CustomersRestApi\Processor\Addresses;
 
+use Generated\Shared\Transfer\AddressesTransfer;
 use Generated\Shared\Transfer\AddressTransfer;
+use Generated\Shared\Transfer\CustomerResponseTransfer;
 use Generated\Shared\Transfer\CustomerTransfer;
 use Generated\Shared\Transfer\RestAddressAttributesTransfer;
 use Generated\Shared\Transfer\RestErrorMessageTransfer;
 use Spryker\Glue\CustomersRestApi\CustomersRestApiConfig;
 use Spryker\Glue\CustomersRestApi\Dependency\Client\CustomersRestApiToCustomerClientInterface;
-use Spryker\Glue\CustomersRestApi\Processor\Customers\CustomersReaderInterface;
 use Spryker\Glue\CustomersRestApi\Processor\Mapper\AddressesResourceMapperInterface;
 use Spryker\Glue\GlueApplication\Rest\JsonApi\RestResourceBuilderInterface;
 use Spryker\Glue\GlueApplication\Rest\JsonApi\RestResponseInterface;
@@ -38,26 +39,18 @@ class AddressesWriter implements AddressesWriterInterface
     protected $addressesResourceMapper;
 
     /**
-     * @var \Spryker\Glue\CustomersRestApi\Processor\Customers\CustomersReaderInterface
-     */
-    protected $customersReader;
-
-    /**
      * @param \Spryker\Glue\GlueApplication\Rest\JsonApi\RestResourceBuilderInterface $restResourceBuilder
      * @param \Spryker\Glue\CustomersRestApi\Dependency\Client\CustomersRestApiToCustomerClientInterface $customerClient
      * @param \Spryker\Glue\CustomersRestApi\Processor\Mapper\AddressesResourceMapperInterface $addressesResourceMapper
-     * @param \Spryker\Glue\CustomersRestApi\Processor\Customers\CustomersReaderInterface $customersReader
      */
     public function __construct(
         RestResourceBuilderInterface $restResourceBuilder,
         CustomersRestApiToCustomerClientInterface $customerClient,
-        AddressesResourceMapperInterface $addressesResourceMapper,
-        CustomersReaderInterface $customersReader
+        AddressesResourceMapperInterface $addressesResourceMapper
     ) {
         $this->restResourceBuilder = $restResourceBuilder;
         $this->customerClient = $customerClient;
         $this->addressesResourceMapper = $addressesResourceMapper;
-        $this->customersReader = $customersReader;
     }
 
     /**
@@ -70,28 +63,31 @@ class AddressesWriter implements AddressesWriterInterface
     {
         $restResponse = $this->restResourceBuilder->createRestResponse();
 
-        $customerReference = $restRequest->findParentResourceByType(CustomersRestApiConfig::RESOURCE_CUSTOMERS)->getId();
+        $customerResponseTransfer = $this->findCustomer($restRequest);
+        $restResponse = $this->validateCustomerResponseTransfer($customerResponseTransfer, $restRequest, $restResponse);
 
-        $customerTransfer = (new CustomerTransfer())
-            ->setCustomerReference($customerReference);
-        $customerResponseTransfer = $this->customerClient->findCustomerByReference($customerTransfer);
-
-        if (!$customerResponseTransfer->getHasCustomer()) {
-            $this->createErrorCustomerNotFound($restResponse);
-
+        if (count($restResponse->getErrors()) > 0) {
             return $restResponse;
         }
 
         $addressTransfer = $this->addressesResourceMapper->mapRestAddressAttributesTransferToAddressTransfer($addressAttributesTransfer);
-        $addressTransfer->setFkCustomer($customerTransfer->getIdCustomer());
+        $addressTransfer->setFkCustomer($customerResponseTransfer->getCustomerTransfer()->getIdCustomer());
 
         $addressTransfer = $this->customerClient->createAddress($addressTransfer);
 
         if (!$addressTransfer->getUuid()) {
-            $this->createErrorAddressNotSaved($restResponse);
-
-            return $restResponse;
+            return $this->createErrorAddressNotSaved($restResponse);
         }
+
+        $addressesTransfer = $this->customerClient->getAddresses($customerResponseTransfer->getCustomerTransfer());
+
+        $addressTransfer = $this->setCustomersDefaultAddresses(
+            $addressTransfer,
+            $addressAttributesTransfer,
+            $addressesTransfer->getAddresses()->count() === 1
+        );
+
+        $this->saveCustomerDefaultAddresses($addressTransfer, $customerResponseTransfer->getCustomerTransfer());
 
         $restResource = $this
             ->addressesResourceMapper
@@ -115,17 +111,31 @@ class AddressesWriter implements AddressesWriterInterface
     {
         $restResponse = $this->restResourceBuilder->createRestResponse();
 
-        $customerReference = $restRequest->findParentResourceByType(CustomersRestApiConfig::RESOURCE_CUSTOMERS)->getId();
-        $customerResponseTransfer = $this->customersReader->findCustomerByReference($customerReference);
+        if (!$restRequest->getResource()->getId()) {
+            return $this->createAddressUuidMissingError($restResponse);
+        }
 
-        $addressTransfer = (new AddressTransfer())
-            ->fromArray($addressAttributesTransfer->toArray(), true);
+        $customerResponseTransfer = $this->findCustomer($restRequest);
+        $restResponse = $this->validateCustomerResponseTransfer($customerResponseTransfer, $restRequest, $restResponse);
+
+        if (count($restResponse->getErrors()) > 0) {
+            return $restResponse;
+        }
+
+        $addressesTransfer = $this->customerClient->getAddresses($customerResponseTransfer->getCustomerTransfer());
+        $addressTransfer = $this->findAddressByUuid($addressesTransfer, $restRequest->getResource()->getId());
+
+        if (!$addressTransfer) {
+            return $this->createAddressNotFoundError($restResponse);
+        }
+
+        $addressTransfer->fromArray($addressAttributesTransfer->modifiedToArray(), true);
+
+        $this->saveCustomerDefaultAddresses($addressTransfer, $customerResponseTransfer->getCustomerTransfer());
         $addressTransfer = $this->customerClient->updateAddress($addressTransfer);
 
         if (!$addressTransfer->getUuid()) {
-            $this->createErrorAddressNotSaved($restResponse);
-
-            return $restResponse;
+            return $this->createErrorAddressNotSaved($restResponse);
         }
 
         $restResource = $this
@@ -149,14 +159,28 @@ class AddressesWriter implements AddressesWriterInterface
     {
         $restResponse = $this->restResourceBuilder->createRestResponse();
 
-        $addressTransfer = (new AddressTransfer())
-            ->setUuid($restRequest->getResource()->getId());
-        $addressTransfer = $this->customerClient->findAddressByUuid($addressTransfer);
+        $customerReference = $restRequest->findParentResourceByType(CustomersRestApiConfig::RESOURCE_CUSTOMERS)->getId();
+
+        if (!$restRequest->getResource()->getId()) {
+            return $this->createAddressUuidMissingError($restResponse);
+        }
+
+        $customerTransfer = (new CustomerTransfer())->setCustomerReference($customerReference);
+        $customerResponseTransfer = $this->customerClient->findCustomerByReference($customerTransfer);
+
+        if (!$customerResponseTransfer->getHasCustomer()) {
+            return $this->createErrorCustomerNotFound($restResponse);
+        }
+
+        if (!$this->isSameCustomerReference($restRequest)) {
+            return $this->createUnauthorizedError($restResponse);
+        }
+
+        $addressesTransfer = $this->customerClient->getAddresses($customerResponseTransfer->getCustomerTransfer());
+        $addressTransfer = $this->findAddressByUuid($addressesTransfer, $restRequest->getResource()->getId());
 
         if (!$addressTransfer) {
-            $this->createAddressNotFoundError($restResponse);
-
-            return $restResponse;
+            return $this->createAddressNotFoundError($restResponse);
         }
 
         $this->customerClient->deleteAddress($addressTransfer);
@@ -207,5 +231,188 @@ class AddressesWriter implements AddressesWriterInterface
             ->setDetail(CustomersRestApiConfig::RESPONSE_DETAILS_ADDRESS_NOT_FOUND);
 
         return $restResponse->addError($restErrorTransfer);
+    }
+
+    /**
+     * @param \Spryker\Glue\GlueApplication\Rest\JsonApi\RestResponseInterface $restResponse
+     *
+     * @return \Spryker\Glue\GlueApplication\Rest\JsonApi\RestResponseInterface
+     */
+    protected function createUnauthorizedError(RestResponseInterface $restResponse): RestResponseInterface
+    {
+        $restErrorTransfer = (new RestErrorMessageTransfer())
+            ->setCode(CustomersRestApiConfig::RESPONSE_CODE_CUSTOMER_UNAUTHORIZED)
+            ->setStatus(Response::HTTP_FORBIDDEN)
+            ->setDetail(CustomersRestApiConfig::RESPONSE_DETAILS_CUSTOMER_UNAUTHORIZED);
+
+        return $restResponse->addError($restErrorTransfer);
+    }
+
+    /**
+     * @param \Spryker\Glue\GlueApplication\Rest\JsonApi\RestResponseInterface $restResponse
+     *
+     * @return \Spryker\Glue\GlueApplication\Rest\JsonApi\RestResponseInterface
+     */
+    protected function createAddressUuidMissingError(RestResponseInterface $restResponse): RestResponseInterface
+    {
+        $restErrorTransfer = (new RestErrorMessageTransfer())
+            ->setStatus(Response::HTTP_BAD_REQUEST)
+            ->setCode(CustomersRestApiConfig::RESPONSE_CODE_ADDRESS_UUID_MISSING)
+            ->setDetail(CustomersRestApiConfig::RESPONSE_DETAILS_ADDRESS_UUID_MISSING);
+
+        return $restResponse->addError($restErrorTransfer);
+    }
+
+    /**
+     * @param \Spryker\Glue\GlueApplication\Rest\Request\Data\RestRequestInterface $restRequest
+     *
+     * @return bool
+     */
+    protected function isSameCustomerReference(RestRequestInterface $restRequest): bool
+    {
+        return $restRequest->getUser()->getNaturalIdentifier()
+            === $restRequest->findParentResourceByType(CustomersRestApiConfig::RESOURCE_CUSTOMERS)->getId();
+    }
+
+    /**
+     * @param \Generated\Shared\Transfer\AddressTransfer $addressTransfer
+     * @param \Generated\Shared\Transfer\CustomerTransfer $customerTransfer
+     *
+     * @return void
+     */
+    protected function saveCustomerDefaultAddresses(
+        AddressTransfer $addressTransfer,
+        CustomerTransfer $customerTransfer
+    ): void {
+        $this->updateCustomerDefaultBillingAddress($addressTransfer, $customerTransfer);
+        $this->updateCustomerDefaultShippingAddress($addressTransfer, $customerTransfer);
+
+        if ($customerTransfer->isPropertyModified(CustomerTransfer::DEFAULT_BILLING_ADDRESS)
+            || $customerTransfer->isPropertyModified(CustomerTransfer::DEFAULT_SHIPPING_ADDRESS)) {
+            $this->customerClient->updateCustomer($customerTransfer);
+        }
+    }
+
+    /**
+     * @param \Generated\Shared\Transfer\AddressTransfer $addressTransfer
+     * @param \Generated\Shared\Transfer\CustomerTransfer $customerTransfer
+     *
+     * @return \Generated\Shared\Transfer\CustomerTransfer
+     */
+    protected function updateCustomerDefaultBillingAddress(
+        AddressTransfer $addressTransfer,
+        CustomerTransfer $customerTransfer
+    ): CustomerTransfer {
+        if ($customerTransfer->getDefaultBillingAddress() === $addressTransfer->getIdCustomerAddress()
+            && $addressTransfer->getIsDefaultBilling() === false
+        ) {
+            $customerTransfer->setDefaultBillingAddress(null);
+        }
+
+        if ($addressTransfer->getIsDefaultBilling() === true
+            && $customerTransfer->getDefaultBillingAddress() !== $addressTransfer->getIdCustomerAddress()
+        ) {
+            $customerTransfer->setDefaultBillingAddress($addressTransfer->getIdCustomerAddress());
+        }
+
+        return $customerTransfer;
+    }
+
+    /**
+     * @param \Generated\Shared\Transfer\AddressTransfer $addressTransfer
+     * @param \Generated\Shared\Transfer\CustomerTransfer $customerTransfer
+     *
+     * @return \Generated\Shared\Transfer\CustomerTransfer
+     */
+    protected function updateCustomerDefaultShippingAddress(
+        AddressTransfer $addressTransfer,
+        CustomerTransfer $customerTransfer
+    ): CustomerTransfer {
+        if ($customerTransfer->getDefaultShippingAddress() === $addressTransfer->getIdCustomerAddress()
+            && $addressTransfer->getIsDefaultShipping() === false
+        ) {
+            $customerTransfer->setDefaultShippingAddress(null);
+        }
+
+        if ($addressTransfer->getIsDefaultShipping() === true
+            && $customerTransfer->getDefaultShippingAddress() !== $addressTransfer->getIdCustomerAddress()
+        ) {
+            $customerTransfer->setDefaultShippingAddress($addressTransfer->getIdCustomerAddress());
+        }
+
+        return $customerTransfer;
+    }
+
+    /**
+     * @param \Generated\Shared\Transfer\AddressesTransfer $addressesTransfer
+     * @param string $uuid
+     *
+     * @return \Generated\Shared\Transfer\AddressTransfer|null
+     */
+    protected function findAddressByUuid(AddressesTransfer $addressesTransfer, string $uuid): ?AddressTransfer
+    {
+        foreach ($addressesTransfer->getAddresses() as $address) {
+            if ($address->getUuid() === $uuid) {
+                return $address;
+            }
+        }
+
+        return null;
+    }
+
+    /**
+     * @param \Generated\Shared\Transfer\AddressTransfer $addressTransfer
+     * @param \Generated\Shared\Transfer\RestAddressAttributesTransfer $addressAttributesTransfer
+     * @param bool $isFirstAddress
+     *
+     * @return \Generated\Shared\Transfer\AddressTransfer
+     */
+    protected function setCustomersDefaultAddresses(AddressTransfer $addressTransfer, RestAddressAttributesTransfer $addressAttributesTransfer, bool $isFirstAddress)
+    {
+        if (!$isFirstAddress) {
+            $addressTransfer->setIsDefaultBilling($addressAttributesTransfer->getIsDefaultBilling());
+            $addressTransfer->setIsDefaultShipping($addressAttributesTransfer->getIsDefaultShipping());
+        } else {
+            $addressTransfer->setIsDefaultBilling(true);
+            $addressTransfer->setIsDefaultShipping(true);
+        }
+
+        return $addressTransfer;
+    }
+
+    /**
+     * @param \Spryker\Glue\GlueApplication\Rest\Request\Data\RestRequestInterface $restRequest
+     *
+     * @return \Generated\Shared\Transfer\CustomerResponseTransfer
+     */
+    protected function findCustomer(RestRequestInterface $restRequest): CustomerResponseTransfer
+    {
+        $customerReference = $restRequest->findParentResourceByType(CustomersRestApiConfig::RESOURCE_CUSTOMERS)->getId();
+        $customerTransfer = (new CustomerTransfer())->setCustomerReference($customerReference);
+
+        return $this->customerClient->findCustomerByReference($customerTransfer);
+    }
+
+    /**
+     * @param \Generated\Shared\Transfer\CustomerResponseTransfer $customerResponseTransfer
+     * @param \Spryker\Glue\GlueApplication\Rest\Request\Data\RestRequestInterface $restRequest
+     * @param \Spryker\Glue\GlueApplication\Rest\JsonApi\RestResponseInterface $restResponse
+     *
+     * @return \Spryker\Glue\GlueApplication\Rest\JsonApi\RestResponseInterface
+     */
+    protected function validateCustomerResponseTransfer(
+        CustomerResponseTransfer $customerResponseTransfer,
+        RestRequestInterface $restRequest,
+        RestResponseInterface $restResponse
+    ): RestResponseInterface {
+        if (!$customerResponseTransfer->getHasCustomer()) {
+            return $this->createErrorCustomerNotFound($restResponse);
+        }
+
+        if (!$this->isSameCustomerReference($restRequest)) {
+            return $this->createUnauthorizedError($restResponse);
+        }
+
+        return $restResponse;
     }
 }
